Move ConnectionListRow into own file
